feat: atom 적용, 구조변경
