+++ src/features/write/ui/calendarForm.tsx
@@ -6,20 +6,10 @@ import DetailCalendar from "./DetailCalendar";
 import { Interview } from "../../../entities/events/model/Interview";
 
 const CalendarForm = ({
-  selectedDate,
-  onDateChange,
-  currentMonth,
-  onMonthChange,
-  events,
   onAddEvent,
   onDeleteEvent,
   onUpdateEvent,
 }: {
-  selectedDate: Date | null;
-  onDateChange: (date: Date | null) => void;
-  currentMonth: Date;
-  onMonthChange: (month: Date) => void;
-  events: { [key: string]: Interview[] };
   onAddEvent: (newEvent: Interview) => void;
   onDeleteEvent: (index: number) => void;
   onUpdateEvent: (index: number, updatedEvent: Interview) => void;
@@ -52,28 +42,13 @@ const CalendarForm = ({
 
   return (
     <Box mx="auto" mt={10} p={4} placeItems="center">
-      <DetailCalendar
-        selectedDate={selectedDate}
-        onDateChange={onDateChange}
-        currentMonth={currentMonth}
-        onMonthChange={onMonthChange}
-        events={events}
-      />
+      <DetailCalendar />
       <EventDialog
         newEvent={newEvent}
         onChange={handleInputChange}
         onAdd={handleAddEvent}
       />
-      <EventManager
-        events={
-          selectedDate
-            ? events[selectedDate.toLocaleDateString("sv-SE")] ?? []
-            : []
-        }
-        onDelete={onDeleteEvent}
-        onUpdate={onUpdateEvent}
-        selectedDate={selectedDate}
-      />
+      <EventManager onDelete={onDeleteEvent} onUpdate={onUpdateEvent} />
     </Box>
   );
 };
